+++ CHANGELOG.md
@@ -14,6 +14,8 @@ UNRELEASED
   https://gitlab.com/gitlab-org/gitaly/merge_requests/164
 - Add support for structured logging via logrus
   https://gitlab.com/gitlab-org/gitaly/merge_requests/163
+- Add support for exposing the Gitaly build information via Prometheus
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/168
 - Set GL_PROTOCOL during SmartHTTP.PostReceivePack
   https://gitlab.com/gitlab-org/gitaly/merge_requests/169